refactor: Перемещение папки screens
