+++ project/src/components/app/app.tsx
@@ -1,4 +1,4 @@
-import MainScreen from '../screens/main-screen/main-screen';
+import MainScreen from '../../screens/main-screen/main-screen';
 
 type AppProps = {
   placesCount: number;
