+++ project/src/screens/favorites-empty-screen/favorites-empty-screen.tsx
