+++ project/src/screens/not-logged-offer-screen/not-logged-offer-screen.tsx
