+++ project/src/screens/offer-screen/offer-screen.tsx
